+++ src/main/java/com/kitisplode/golemfirststonemod/item/ModItemGroups.java
@@ -26,8 +26,10 @@ public class ModItemGroups
                         entries.add(ModBlocks.BLOCK_CORE_BRICK);
                         entries.add(ModBlocks.BLOCK_HEAD_DIORITE);
                         entries.add(ModBlocks.BLOCK_CORE_DIORITE);
+                        entries.add(ModBlocks.BLOCK_DANDORI);
                         entries.add(ModItems.ITEM_SPAWN_PAWN_FIRST_DIORITE);
                         entries.add(ModItems.ITEM_DANDORI_CALL);
+                        entries.add(ModItems.ITEM_DANDORI_ATTACK);
                     }).build());
 
     public static void registerItemGroups()