+++ src/main/java/com/kitisplode/golemfirststonemod/sound/ModSounds.java
@@ -9,6 +9,8 @@ import net.minecraft.util.Identifier;
 public class ModSounds
 {
     public static SoundEvent ITEM_DANDORI_CALL = registerSoundEvent("item_dandori_call");
+    public static SoundEvent ITEM_DANDORI_ATTACK_FAIL = registerSoundEvent("item_dandori_attack_fail");
+    public static SoundEvent ITEM_DANDORI_ATTACK_WIN = registerSoundEvent("item_dandori_attack_win");
 
     private static SoundEvent registerSoundEvent(String name)
     {