+++ CHANGELOG.md
@@ -22,3 +22,14 @@
     + added Diorite Pawn entity
     + added Sculpted Gold Head and Chiseled Diorite Core blocks
     + added First of Diorite structure
+
+## 0.0.5
+    + added Dandori system
+    + added Golem Dandori Call item
+    + integrated Dandori system into Iron and Snow Golems
+    + added Dandori Villager profession
+    + QoL - Made AoE attacks not damage a player's ride
+    + QoL - Made AoE shields apply to a player's ride
+    + Balance - Made Diorite Pawns stay closer to their First of Diorite
+    + Balance - Made Diorite Pawns die off if they don't have a target for 30 seconds
+    + Balance - Reduced number of Diorite Pawns spawned to 3 at a time